State na home. props pra filtro e card
